requirements.txt
uv 
ddgs
beautifulsoup4 
requests
lxml
duckduckgo_search
openai
tavily
cloudinary
python-multipart
cloudinary
wikipedia
openai 
pillow
python-dotenv
google-genai
anthropic